Added LFO app

+++ midipal/apps/active_sensing_filter.h
@@ -32,7 +32,7 @@ class ActiveSensingFilter : public App {
   void OnRawByte(uint8_t byte);
   
   uint8_t settings_size() { return 1; }
-  uint8_t settings_offset() { return SETTINGS_0XFE_FILTER; }
+  uint16_t settings_offset() { return SETTINGS_0XFE_FILTER; }
   uint8_t* settings_data() { return &enabled_; }
   uint8_t app_name() { return STR_RES_0XFE_FLT; }
