+++ midipal/apps/dispatcher.h
@@ -56,7 +56,7 @@ class Dispatcher : public App {
      uint8_t velocity);
  
   uint8_t settings_size() { return 4; }
-  uint8_t settings_offset() { return SETTINGS_DISPATCHER; }
+  uint16_t settings_offset() { return SETTINGS_DISPATCHER; }
   uint8_t* settings_data() { return &input_channel_; }
   uint8_t app_name() { return STR_RES_DISPATCH; }
