+++ midipal/apps/drum_pattern_generator.h
@@ -48,7 +48,7 @@ class DrumPatternGenerator : public App {
   void SetParameter(uint8_t key, uint8_t value);
   
   uint8_t settings_size() { return 10; }
-  uint8_t settings_offset() { return SETTINGS_DRUM_PATTERN_GENERATOR; }
+  uint16_t settings_offset() { return SETTINGS_DRUM_PATTERN_GENERATOR; }
   uint8_t* settings_data() { return &mode_; }
   uint8_t app_name() { return STR_RES_DRUMS; }
