+++ midipal/apps/ear_training_game.h
@@ -44,7 +44,7 @@ class EarTrainingGame : public App {
   uint8_t OnIncrement(int8_t value);
   
   uint8_t settings_size() { return 7; }
-  uint8_t settings_offset() { return SETTINGS_EAR_TRAINING; }
+  uint16_t settings_offset() { return SETTINGS_EAR_TRAINING; }
   uint8_t* settings_data() { return &level_; }
   uint8_t app_name() { return STR_RES_EAR_GAME; }
